+++ src/eggviron/__init__.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 
-from .eggviron import Eggviron
-from .envfile_loader import EnvFileLoader
-from .environ_loader import EnvironLoader
+from ._eggviron import Eggviron
+from ._envfile_loader import EnvFileLoader
+from ._environ_loader import EnvironLoader
 
 __all__ = [
     "Eggviron",
